// types.ts
export enum ModuleType {
  DASHBOARD = 'DASHBOARD',
  HR = 'HR',
  ENGINEERING = 'ENGINEERING',
  TIME_TRACKING = 'TIME_TRACKING',
  TASKS = 'TASKS',
  CHAT = 'CHAT',
  ADMIN_SETTINGS = 'ADMIN_SETTINGS'
}

export type ChatType = 'DIRECT' | 'GROUP' | 'CHANNEL_NEWS' | 'CHANNEL_DEPT' | 'CHANNEL_ORDERS';

export interface UserPermissions {
  [ModuleType.DASHBOARD]: boolean;
  [ModuleType.HR]: boolean;
  [ModuleType.ENGINEERING]: boolean;
  [ModuleType.TIME_TRACKING]: boolean;
  [ModuleType.TASKS]: boolean;
  [ModuleType.CHAT]: boolean;
}

export interface User {
  id: string;
  login: string;
  password?: string;
  fullName: string;
  role: 'ADMIN' | 'USER';
  permissions: UserPermissions;
}